Im keeping with the commit early - commit often mantra here, the code is not really ready :) Since the code had become increasingly slow, i did some optimizations (and of course caused some breakage) - use ints with two decimal points worth of precision instead of floats    this should also simplify some code once im done - wrote my own Point and Rectangle classes, as the ones i used only partly fit (and there were no integer-only alternatives)    took advantage of this to clean up output, though its also not done yet, its somewhat excessive - avoid needless object allocations, obtained quite a speed up in some code paths - some things are moved around a bit in TextExtractor, but im not sure if those changes are final, its quite messy now - optimized the code for writing tree to file

+++ src/main/java/org/elacin/pdfextract/tree/ParagraphNode.java
@@ -142,12 +142,12 @@ public class ParagraphNode extends AbstractParentNode<LineNode, PageNode> {
         /* Use the size of the preceding and new text to determine if the two lines are
              logically part of the same paragraph. Use the smaller of the two values.
          */
-        double ySize = Math.min(lastLineNode.getCurrentStyle().ySize, node.getStyle().ySize);
-        //        double ySize = lastLineNode.getCurrentStyle().ySize;
+        //TODO: style!
+        int ySize = Math.min(lastLineNode.getCurrentStyle().ySize, node.getStyle().ySize);
 
         if (isWithinVariance(getPosition().getY() + getPosition().getHeight(), node.getPosition().getY(), ySize)) {
 
-            double widthOfSpace = lastLineNode.getCurrentStyle().widthOfSpace;
+            int widthOfSpace = lastLineNode.getCurrentStyle().widthOfSpace;
 
             /* if the new node STARTS at more or less the same X position, accept it */
             if (isWithinVariance(getPosition().getX(), node.getPosition().getX(), widthOfSpace)) {
